Initial tests for babel-register
