+++ packages/babel-register/test/__data__/es2015.js
@@ -1,0 +1,3 @@
+import a from "assert";
+
+export default () => { a.foo(); return "b"; };
