+++ packages/babel-register/test/index.js
@@ -1,84 +1,45 @@
-import { expect } from "chai";
-import fs from "fs";
+import assert from "assert";
 import path from "path";
 import decache from "decache";
 
-const testCacheFilename = path.join(__dirname, ".babel");
-const oldBabelDisableCacheValue = process.env.BABEL_DISABLE_CACHE;
+const DATA_ES2015 = require.resolve("./__data__/es2015");
 
-process.env.BABEL_CACHE_PATH = testCacheFilename;
-delete process.env.BABEL_DISABLE_CACHE;
+describe("babel-register", function () {
+  let babelRegister = null;
 
-function writeCache(data) {
-  if (typeof data === "object") {
-    data = JSON.stringify(data);
+  function setupRegister(config) {
+    babelRegister = require("../lib/node");
+    babelRegister.default(Object.assign({
+      presets: [path.join(__dirname, "../../babel-preset-es2015")],
+      babelrc: false,
+      only: ["__data__"],
+    }, config));
   }
 
-  fs.writeFileSync(testCacheFilename, data);
-}
-
-function cleanCache() {
-
-  try {
-    fs.unlinkSync(testCacheFilename);
-  } catch (e) {
-    // It is convenient to always try to clear
+  function revertRegister() {
+    if (babelRegister) {
+      babelRegister.revert();
+      babelRegister = null;
+    }
   }
-}
 
-function resetCache() {
-  process.env.BABEL_CACHE_PATH = null;
-  process.env.BABEL_DISABLE_CACHE = oldBabelDisableCacheValue;
-}
+  afterEach(() => {
+    revertRegister();
+    decache(DATA_ES2015);
+  });
 
-describe("babel register", () => {
+  it("basic usage", function () {
+    setupRegister();
 
-  describe("cache", () => {
-    let load, get, save;
+    assert.ok(require(DATA_ES2015).default);
+  });
 
-    beforeEach(() => {
-      // Since lib/cache is a singleton we need to fully reload it
-      decache("../lib/cache");
-      const cache = require("../lib/cache");
+  it("reverts correctly", () => {
+    setupRegister();
+    revertRegister();
 
-      load = cache.load;
-      get = cache.get;
-      save = cache.save;
-    });
-
-    afterEach(cleanCache);
-    after(resetCache);
-
-    it("should load and get cached data", () => {
-      writeCache({ foo: "bar" });
-
-      load();
-
-      expect(get()).to.be.an("object");
-      expect(get()).to.deep.equal({ foo: "bar" });
-    });
-
-    it("should load and get an object with no cached data", () => {
-      load();
-
-      expect(get()).to.be.an("object");
-      expect(get()).to.deep.equal({});
-    });
-
-    it("should load and get an object with invalid cached data", () => {
-      writeCache("foobar");
-
-      load();
-
-      expect(get()).to.be.an("object");
-      expect(get()).to.deep.equal({});
-    });
-
-    it("should create the cache on save", () => {
-      save();
-
-      expect(fs.existsSync(testCacheFilename)).to.be.true;
-      expect(get()).to.deep.equal({});
-    });
+    assert.throws(function () {
+      require(DATA_ES2015);
+    }, SyntaxError);
   });
 });
